requirement.txt
certifi==2017.11.5
chardet==3.0.4
idna==2.6
lxml==4.1.1
olefile==0.44
Pillow==4.3.0
requests==2.18.4
urllib3==1.22
